@@ -17,14 +17,7 @@
 
 // Functionality for API import/export symbols, in a platform independent way.
 // When building the DLL, use the ChApiEXPORT macro. When using the DLL, use the ChApiIMPORT macro.
-
-#if ((defined _WIN32) || (defined(__MINGW32__) || defined(__CYGWIN__))) && defined(_DLL)
-#if !defined(CH_DLL) && !defined(CH_STATIC)
-#define CH_DLL
-#endif
-#endif
-
-#if ((defined _WIN32) || (defined(__MINGW32__) || defined(__CYGWIN__))) && defined(CH_DLL)
+#if ((defined _WIN32) || (defined(__MINGW32__) || defined(__CYGWIN__))) && !defined(CH_STATIC)
 #define ChApiEXPORT __declspec(dllexport)
 #define ChApiIMPORT __declspec(dllimport)
 #else
