@@ -69,7 +69,7 @@ class ChApi CSV_writer {
 
     void write_to_file(const std::string& filename, const std::string& header = "") const {
         std::ofstream ofile(filename.c_str());
-        ofile << header;
+        ofile << header << std::endl;
         ofile << m_ss.str();
         ofile.close();
     }
