@@ -107,31 +107,30 @@ int main() {
 
     // Scenario: driven wheel
     ////rig.SetAngSpeedFunction(chrono_types::make_shared<ChFunction_Const>(10.0));
-    ////rig.Initialize();
 
     // Scenario: pulled wheel
     ////rig.SetLongSpeedFunction(chrono_types::make_shared<ChFunction_Const>(1.0));
-    ////rig.Initialize();
 
-    // Scenario: imobilized wheel
+    // Scenario: imobilized wheel (same scenario could be obtained using ChTireTestRig::Mode::DROP in Initialize())
     rig.SetLongSpeedFunction(chrono_types::make_shared<ChFunction_Const>(0.0));
     rig.SetAngSpeedFunction(chrono_types::make_shared<ChFunction_Const>(0.0));
-    rig.Initialize();
 
     // Scenario: prescribe all motion functions
     ////rig.SetLongSpeedFunction(chrono_types::make_shared<ChFunction_Const>(0.2));
     ////rig.SetAngSpeedFunction(chrono_types::make_shared<ChFunction_Const>(10.0));
     ////rig.SetSlipAngleFunction(chrono_types::make_shared<ChFunction_Sine>(0, 0.6, 0.2));
-    ////rig.Initialize();
 
     // Scenario: specified longitudinal slip
-    ////rig.Initialize(0.2, 1.0);
+    ////rig.SetConstantLongitudinalSlip(0.2, 0.1);
+
+    // Initialize the tire test rig
+    rig.SetTimeDelay(1.0);
+    rig.Initialize(ChTireTestRig::Mode::TEST);
 
     // Display settings
     // ----------------
 
     std::cout << "Total rig mass: " << rig.GetMass() << std::endl;
-    std::cout << "Applied load:   " << rig.GetAppliedLoad() << std::endl;
 
     // Initialize OpenGL
     // -----------------
